# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ante",
    "description": "AI-Native Trading Engine API",
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/portfolio/value": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Portfolio Value",
        "description": "총 자산 가치 + 당일 손익 + 수익률 + 미실현 손익 (최신 스냅샷 기반).\n인증된 master/human 또는 ``treasury:read`` scope 를 보유한 agent 만 호출\n가능 (#1407 — portfolio 라우트는 ``treasury_daily_snapshots`` 를 읽으므로\nspec ``treasury:read`` 정합).",
        "operationId": "portfolio_value_api_portfolio_value_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioValueResponse"
                }
              }
            }
          },
          "503": {
            "description": "Treasury not available",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/history": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Portfolio History",
        "description": "기간별 자산 추이 (daily_snapshots 시계열 반환). 인증된 master/human 또는\n``treasury:read`` scope 를 보유한 agent 만 호출 가능 (#1407).",
        "operationId": "portfolio_history_api_portfolio_history_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioHistoryResponse"
                }
              }
            }
          },
          "503": {
            "description": "Treasury not available",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortfolioHistoryPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "total_asset": {
            "type": "number",
            "title": "Total Asset"
          },
          "daily_pnl": {
            "type": "number",
            "title": "Daily Pnl"
          },
          "daily_return": {
            "type": "number",
            "title": "Daily Return"
          },
          "unrealized_pnl": {
            "type": "number",
            "title": "Unrealized Pnl"
          }
        },
        "type": "object",
        "required": [
          "date",
          "total_asset",
          "daily_pnl",
          "daily_return",
          "unrealized_pnl"
        ],
        "title": "PortfolioHistoryPoint",
        "description": "자산 추이 데이터 포인트 (스냅샷 기반)."
      },
      "PortfolioHistoryResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PortfolioHistoryPoint"
            },
            "type": "array",
            "title": "Data"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "data",
          "start_date",
          "end_date"
        ],
        "title": "PortfolioHistoryResponse",
        "description": "기간별 자산 추이 응답."
      },
      "PortfolioValueResponse": {
        "properties": {
          "total_value": {
            "type": "number",
            "title": "Total Value"
          },
          "daily_pnl": {
            "type": "number",
            "title": "Daily Pnl"
          },
          "daily_return": {
            "type": "number",
            "title": "Daily Return"
          },
          "unrealized_pnl": {
            "type": "number",
            "title": "Unrealized Pnl",
            "default": 0.0
          },
          "snapshot_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Date"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "total_value",
          "daily_pnl",
          "daily_return",
          "updated_at"
        ],
        "title": "PortfolioValueResponse",
        "description": "총 자산 가치 응답 (스냅샷 기반)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ErrorResponse": {
        "description": "RFC 7807 Problem Details 에러 응답.",
        "properties": {
          "type": {
            "default": "/errors/internal",
            "title": "Type",
            "type": "string"
          },
          "title": {
            "default": "Internal Server Error",
            "title": "Title",
            "type": "string"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "status": {
            "default": 500,
            "title": "Status",
            "type": "integer"
          },
          "instance": {
            "default": "",
            "title": "Instance",
            "type": "string"
          }
        },
        "title": "ErrorResponse",
        "type": "object"
      },
...
    }
  }
}